+++ lib/src/parse/stylesheet.dart
@@ -2289,7 +2289,9 @@ abstract class StylesheetParser extends Parser {
 
         case $exclamation:
         case $semicolon:
-          break loop;
+          if (brackets.isEmpty) break loop;
+          buffer.writeCharCode(scanner.readChar());
+          break;
 
         case $u:
         case $U: